Expand alloc features further
